refactor: Move DefaultEmissionsService to internal
