+++ application/src/ext/java/org/opentripplanner/ext/emissions/configure/EmissionsServiceModule.java
@@ -3,9 +3,9 @@ package org.opentripplanner.ext.emissions.configure;
 import dagger.Module;
 import dagger.Provides;
 import jakarta.inject.Singleton;
-import org.opentripplanner.ext.emissions.DefaultEmissionsService;
 import org.opentripplanner.ext.emissions.EmissionsDataModel;
 import org.opentripplanner.ext.emissions.EmissionsService;
+import org.opentripplanner.ext.emissions.internal.DefaultEmissionsService;
 
 /**
  * The service is used during application serve phase, not loading, so we need to provide
